Preload compliment images after background image
As soon as we’ve added the background image, instruct the browser to
preload the compliment images, since we’ll need (some of) them soon.
(There’s no need to preload audio – the browser does that as soon as we
create the new Audio(), which happens as soon as the compliment is
loaded, not when the “play” button is clicked, so there’s enough time
for preloading to happen until we need to play the audio.)

Note that Firefox currently doesn’t support rel=preload, but that’s
fine: it just gracefully degrades to on-demand loading.