Scroll whole page, not just <article>
According to the spec, in case of overflow the whole page is supposed to
be scrollable, not just the article with the compliment. (I suppose it
looks better anyways to have the scrollbar only at the outermost level.)
This requires a solution to the “flex causes overflow” problem, so
horizontally center the <article> differently, by adding margin-inline:
auto to the body (and limiting the body’s inline-size so it doesn’t just
fill the whole screen and have the <article> left-aligned inside).

The only unfortunate side effect of this is that if the size of the page
changes (a compliment is loaded that causes it to become scrollable),
the background now jumps around, because it is centered and therefore
moves with the center of the page, not of the screen.