Move background-image to a dedicated <div>
This saves the problem mentioned in the last commit message, where the
background image jumps around if the page size changes due to larger
compliments: we add a <div> that is always as high and wide as the
viewport, and positioned at 0/0, and give the background-image to that
rather than the <body>.

Sadly, this requires us to use the non-logical properties width and
height. Using the 100vh and 100vw units with the logical properties
wouldn’t make sense, and we can’t quite use the direction-independent
100% instead, because that gets affected by the scrollbar if one exists
(so the background-image shifts around slightly as the scrollbar appears
and disappears).