Load background image more efficiently
When loading the image, load its width and height as well. In the HTML,
turn the #background from a <div> into an <img>, which lets us use the
width and height attributes to inform the browser of the intrinsic size
of the image (which we already know from the same API call that also
gives us the attribution – loading the background image is now delayed
until after that API call). And with this, we can also set a srcset on
the image, and given the intrinsic size of the image the browser can
correctly determine which scaled version of the image it should load.
(I’m not sure if browsers actually use this at the moment, but I don’t
see why they shouldn’t.)

The alternate image sizes are taken from [1] – those are the resolutions
which are likely to already exist on the Wikimedia servers.

[1]: https://discourse-mediawiki.wmflabs.org/t/embedding-wikimedia-commons-images-in-tools-with-multiple-resolutions-srcset/1501/2