Fix srcset of background image
Compare the previous commit – if we’re using “width” descriptors, then
the main source must also be included with its width. (I’m not sure why
the browser can’t infer it from the width attribute, in fact, but
whatever, let’s add it.) Without this, Firefox never loaded the
full-size version of the image, it would stop at the 1920px version.