Add image sizes and downscaled versions
Add downscaled versions of each image until the downscaled version would
be less than 608px wide (the effective max-width of the image). The
downscaling was done with `convert FILE.jpg -resize 50% FILE-50.jpg` or
equivalent for 25%. Firefox and Chrome don’t seem to use the smaller
version from the srcset a lot of the time, not sure why, but sometimes
they do, and as far as I can tell nothing’s wrong with the srcset.